requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.0
tqdm==4.65.0
scikit-learn-intelex==2024.4.0
git-python